Restructure project modules
Restructure project modules for convenience
