+++ backend/data/db/storages/currencies.py
