+++ backend/data/managers/models/currencies.py
@@ -1,8 +1,8 @@
 from flask_smorest import abort
 from sqlalchemy.exc import IntegrityError
 
-from backend.models.db import db
-from backend.models.currencies import CurrencyModel
+from backend.data.db.models.db import db
+from backend.data.db.models.currencies import CurrencyModel
 
 
 class CurrenciesManagerORM:
