+++ backend/views/resources/users.py
@@ -2,10 +2,10 @@ from flask.views import MethodView
 from flask_jwt_extended import jwt_required
 from flask_smorest import Blueprint
 
-from backend.schemas.schemas import UserSchema, UserCurrencySchema, UserLoginSchema, UserTokenSchema
+from backend.views.schemas.schemas import UserSchema, UserCurrencySchema, UserLoginSchema, UserTokenSchema
 
-from backend.managers.storages.managers import users_storage
-from backend.managers.models.managers import users_orm, login_orm
+from backend.data.managers.storages.managers import users_storage
+from backend.data.managers.models.managers import users_orm, login_orm
 
 blp = Blueprint(
     "user", __name__, description="Blueprint for operations on users"
